+++ src/dbt_score/__init__.py
@@ -1,11 +1,12 @@
 """Init dbt_score package."""
 
-from dbt_score.models import Exposure, Model, Seed, Snapshot, Source
+from dbt_score.models import Exposure, Macro, Model, Seed, Snapshot, Source
 from dbt_score.rule import Rule, RuleViolation, Severity, rule
 from dbt_score.rule_filter import RuleFilter, rule_filter
 
 __all__ = [
     "Exposure",
+    "Macro",
     "Model",
     "Source",
     "Snapshot",